Fixed warnings caused by wrong casts when using unsigned bit shift operations

+++ Core/Darkorbit/Commands/PostHandshake/ShipRepaired.cs
@@ -19,13 +19,13 @@ namespace PolskaBot.Core.Darkorbit.Commands.PostHandshake
         public ShipRepaired(EndianBinaryReader reader)
         {
             HP = reader.ReadInt32();
-            HP = (int)(HP << 5 | (uint)HP >> 27);
+            HP = (int)((uint)HP << 5 | (uint)HP >> 27);
             MaxHP = reader.ReadInt32();
-            MaxHP = (int)(MaxHP << 8 | (uint)MaxHP >> 24);
+            MaxHP = (int)((uint)MaxHP << 8 | (uint)MaxHP >> 24);
             NanoHP = reader.ReadInt32();
-            NanoHP = (int)(NanoHP << 10 | (uint)NanoHP >> 22);
+            NanoHP = (int)((uint)NanoHP << 10 | (uint)NanoHP >> 22);
             MaxNanoHP = reader.ReadInt32();
-            MaxNanoHP = (int)((uint)MaxNanoHP >> 8 | MaxNanoHP << 24);
+            MaxNanoHP = (int)((uint)MaxNanoHP >> 8 | (uint)MaxNanoHP << 24);
         }
 
         public override void Write()
